+++ test/pthread/test_pthread_unhandledrejection.c
@@ -1,0 +1,6 @@
+#include <emscripten/emscripten.h>
+
+int main() {
+  EM_ASM({ Promise.reject("rejected!"); });
+  emscripten_exit_with_live_runtime();
+}