+++ ChangeLog.md
@@ -24,6 +24,8 @@ See docs/process.md for more on how version tagging works.
   emscripten (llvm, wabt, binaryen).  This should not effect any users of
   emscripten, only developers. (#17502)
 - The llvm version that emscripten uses was updated to 16.0.0 (#17534)
+- worker.js now propagates unhandled promise rejections to the main thread the
+  same way it propagates uncaught exceptions.
 
 3.1.17 - 07/22/2022
 ------